fix(strimzi): add Hadoop + AWS JARs to KafkaConnect, fix SQL type casts
- Add iceberg-aws-bundle, hadoop-client-runtime, and hadoop-client-api
  to KafkaConnect Docker image via Dockerfile.connect — fixes
  NoClassDefFoundError for HadoopCatalog + enables S3FileIO
- Switch warehouse URI from s3a:// to s3:// for S3FileIO compatibility
- Add iceberg.catalog.s3.region to both sink patches
- Fix asyncpg AmbiguousFunctionError in upsert_position by adding
  explicit ::numeric casts (36/36 tests passing)
- Update localdev image reference to data-lake-connect:with-aws

+++ producer-api/app/database.py
@@ -130,11 +130,11 @@ class TradingDatabase:
             INSERT INTO positions (
                 account_id, instrument_id, quantity, market_value, position_date
             )
-            VALUES ($1, $2, $3, $3 * $4, CURRENT_DATE)
+            VALUES ($1, $2, $3, $3::numeric * $4::numeric, CURRENT_DATE)
             ON CONFLICT (account_id, instrument_id, position_date)
             DO UPDATE SET
-                quantity      = positions.quantity + $3,
-                market_value  = (positions.quantity + $3) * $4,
+                quantity      = positions.quantity + $3::numeric,
+                market_value  = (positions.quantity + $3::numeric) * $4::numeric,
                 updated_at    = NOW()
             """,
             account_id,
